Validate and Cleaning up pt1

+++ products/delete-cart.php
@@ -3,6 +3,16 @@
 <?php
 
 
+    if(!isset($_SERVER['HTTP_REFERER'])){
+        // redirect them to your desired location
+        header('location: http://localhost/coffee-blend/index.php');
+        exit;
+    }
+
+    if(!isset($_SESSION['user_id'])){
+        header("location: ".APPURL."");
+    }
+
     $deleteAll = $conn->query("DELETE FROM cart WHERE user_id = '$_SESSION[user_id]'");
     $deleteAll->execute();
